pyproject.toml
[project]
name = "helion-autoresearch"
version = "0.1.0"
description = "Autonomous Helion kernel optimization via AI agents"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pyyaml",
    "torch",
    "helion",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends._legacy:_Backend"
